addon-resizer: Fix --poll-period is not working in 1.8.13
In Addon-resizer 1.8.13 or earlier, there is a problem that
`--poll-period` flag does not change the poll period.

This is caused by a bug while parsing the flag; pollPeriod value
from `flag.Int` is dereferenced too early (before `flag.Parse()`)
and it always be assigned the default value.

This commit resolves this issue by getting the pollPeriod value after
`flag.Parse()`.